+++ auth/__init__.py
@@ -1,0 +1,3 @@
+from .accounts import AccountManager, Account
+
+__all__ = ["AccountManager", "Account"]